requirements.txt
blinker==1.9.0
click==8.3.1
dnspython==2.8.0
Flask==3.0.0
gunicorn==24.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pymongo==4.6.1
python-dateutil==2.9.0
python-dotenv==1.0.1
six==1.17.0
Werkzeug==3.1.5
